refactor(langy): split langy.ts into per-concern route files
langy.ts had grown to 1,800+ lines mixing six concerns: chat streaming,
17 tools, conversation CRUD, project memory CRUD + refresh streaming,
preferences, and privacy (clear-all + GDPR export). Reviewers had to
load the whole file to touch any one of them.

Split the non-chat routes into sibling files, each exporting a
register*Routes(app) function that mutates the shared Hono app:
  - langy.helpers.ts        shared session/admin guards
  - langy.conversations.ts  GET/PATCH/DELETE conversation CRUD
  - langy.project-memory.ts GET/PUT memory + POST refresh stream
  - langy.preferences.ts    GET/PUT user preferences
  - langy.privacy.ts        DELETE memory + GET export

langy.ts is now just the chat stream + 17 tools + the workbench-state
helpers used by tools. Same middleware stack covers every route since
all registrations run on the same Hono instance.

Drive-by: tighten the requireSessionAndPermission return type so callers
no longer need `guard.session!` non-null assertions, and LangyContext
wrapper added to LangyConversationHistory.integration.test.tsx (same
unrelated pre-existing breakage as LangySidebar tests — 16 tests now
pass that were failing before).
