+++ langwatch/src/server/routes/langy.ts
@@ -32,7 +32,6 @@ import { getVercelAIModel } from "~/server/modelProviders/utils";
 import { PromptService } from "~/server/prompt-config/prompt.service";
 import { parseEvaluationResult } from "~/utils/evaluationResults";
 import { createLogger } from "~/utils/logger/server";
-import { auditLog } from "~/server/auditLog";
 import { TiktokenClient } from "~/server/app-layer/clients/tokenizer/tiktoken.client";
 import { buildLangyTelemetrySettings } from "~/server/observability/langy-tracer";
 import {
@@ -41,16 +40,17 @@ import {
   LangyProjectMemoryService,
   LangyUserPreferencesService,
 } from "~/server/services/langy";
-import {
-  LANGY_SYSTEM_PROMPT,
-  PROJECT_MEMORY_REFRESH_PROMPT,
-} from "~/server/services/langy/prompts";
+import { LANGY_SYSTEM_PROMPT } from "~/server/services/langy/prompts";
 import { ConversationToolIdSet } from "~/server/services/langy/toolIdValidator";
 import {
   LANGY_TOOL_CALLS_PER_MESSAGE,
   checkLangyMessageRateLimit,
 } from "~/server/middleware/rate-limit-langy";
 import { esClient, TRACE_INDEX } from "~/server/elasticsearch";
+import { registerLangyConversationRoutes } from "./langy.conversations";
+import { registerLangyProjectMemoryRoutes } from "./langy.project-memory";
+import { registerLangyPreferencesRoutes } from "./langy.preferences";
+import { registerLangyPrivacyRoutes } from "./langy.privacy";
 import type { NextRequestShim as any } from "./types";
 
 const logger = createLogger("langwatch:api:langy");
@@ -1227,366 +1227,14 @@ app.post("/langy/chat", async (c) => {
   });
 });
 
-// ============================================================================
-// Conversation management
-// ============================================================================
-
-async function requireSessionAndPermission(c: any, projectId: string | undefined) {
-  const session = await getServerAuthSession({ req: c.req.raw as any });
-  if (!session) return { error: c.json({ error: "Unauthorized" }, { status: 401 }) };
-  if (!projectId) return { error: c.json({ error: "Missing projectId" }, { status: 400 }) };
-  const ok = await hasProjectPermission(
-    { prisma, session },
-    projectId,
-    "evaluations:view",
-  );
-  if (!ok) return { error: c.json({ error: "Forbidden" }, { status: 403 }) };
-  return { session };
-}
-
-async function requireProjectAdmin(session: Awaited<ReturnType<typeof getServerAuthSession>>, projectId: string) {
-  if (!session) return false;
-  return await hasProjectPermission(
-    { prisma, session },
-    projectId,
-    "project:manage",
-  );
-}
-
-app.get("/langy/conversations", async (c) => {
-  const projectId = c.req.query("projectId");
-  const guard = await requireSessionAndPermission(c, projectId);
-  if (guard.error) return guard.error;
-  const limit = Number(c.req.query("limit") ?? "50");
-  const service = LangyConversationService.create(prisma);
-  const conversations = await service.getAll({
-    projectId: projectId!,
-    userId: guard.session!.user.id,
-    limit: Math.min(Math.max(limit, 1), 100),
-  });
-  return c.json({ conversations });
-});
-
-app.get("/langy/conversations/:id", async (c) => {
-  const projectId = c.req.query("projectId");
-  const guard = await requireSessionAndPermission(c, projectId);
-  if (guard.error) return guard.error;
-  const id = c.req.param("id");
-  const convService = LangyConversationService.create(prisma);
-  const conv = await convService.getById({
-    id,
-    projectId: projectId!,
-    userId: guard.session!.user.id,
-  });
-  if (!conv) return c.json({ error: "Not found" }, { status: 404 });
-  const msgService = LangyMessageService.create(prisma);
-  const messages = await msgService.getAllByConversation({
-    conversationId: conv.id,
-    projectId: projectId!,
-  });
-  return c.json({ conversation: conv, messages });
-});
-
-app.patch("/langy/conversations/:id", async (c) => {
-  const body = (await c.req.json()) as {
-    projectId: string;
-    title?: string | null;
-    isShared?: boolean;
-  };
-  const guard = await requireSessionAndPermission(c, body.projectId);
-  if (guard.error) return guard.error;
-  const id = c.req.param("id");
-  const service = LangyConversationService.create(prisma);
-  try {
-    const updated = await service.updateById({
-      id,
-      projectId: body.projectId,
-      userId: guard.session!.user.id,
-      title: body.title,
-      isShared: body.isShared,
-    });
-    if (body.isShared !== undefined) {
-      await auditLog({
-        userId: guard.session!.user.id,
-        projectId: body.projectId,
-        action: body.isShared
-          ? "langy.conversation.share"
-          : "langy.conversation.unshare",
-        args: { conversationId: id },
-      });
-    }
-    return c.json({ conversation: updated });
-  } catch {
-    return c.json({ error: "Not found or not owned" }, { status: 404 });
-  }
-});
-
-app.delete("/langy/conversations/:id", async (c) => {
-  const projectId = c.req.query("projectId");
-  const guard = await requireSessionAndPermission(c, projectId);
-  if (guard.error) return guard.error;
-  const id = c.req.param("id");
-  const service = LangyConversationService.create(prisma);
-  const ok = await service.deleteById({
-    id,
-    projectId: projectId!,
-    userId: guard.session!.user.id,
-  });
-  if (!ok) return c.json({ error: "Not found or not owned" }, { status: 404 });
-  return c.json({ success: true });
-});
-
-// ============================================================================
-// Project memory
-// ============================================================================
-
-app.get("/langy/project-memory", async (c) => {
-  const projectId = c.req.query("projectId");
-  const guard = await requireSessionAndPermission(c, projectId);
-  if (guard.error) return guard.error;
-  const service = LangyProjectMemoryService.create(prisma);
-  const memory = await service.getById({ projectId: projectId! });
-  const isStale = memory
-    ? await service.isStale({ projectId: projectId! })
-    : false;
-  return c.json({ memory, isStale });
-});
-
-app.put("/langy/project-memory", async (c) => {
-  const body = (await c.req.json()) as { projectId: string; content: string };
-  const guard = await requireSessionAndPermission(c, body.projectId);
-  if (guard.error) return guard.error;
-  const isAdmin = await requireProjectAdmin(guard.session!, body.projectId);
-  if (!isAdmin) {
-    return c.json(
-      { error: "Editing project memory requires project admin." },
-      { status: 403 },
-    );
-  }
-  const service = LangyProjectMemoryService.create(prisma);
-  const memory = await service.writeNewVersion({
-    projectId: body.projectId,
-    content: body.content,
-    changedById: guard.session!.user.id,
-    changeReason: "user_edit",
-  });
-  await auditLog({
-    userId: guard.session!.user.id,
-    projectId: body.projectId,
-    action: "langy.project_memory.edit",
-    args: { contentVersion: memory.contentVersion },
-  });
-  return c.json({ memory });
-});
-
-app.post("/langy/project-memory/refresh", async (c) => {
-  const body = (await c.req.json()) as { projectId: string };
-  const guard = await requireSessionAndPermission(c, body.projectId);
-  if (guard.error) return guard.error;
-  const isAdmin = await requireProjectAdmin(guard.session!, body.projectId);
-  if (!isAdmin) {
-    return c.json(
-      { error: "Refreshing project memory requires project admin." },
-      { status: 403 },
-    );
-  }
-
-  let model;
-  try {
-    model = await getVercelAIModel(body.projectId);
-  } catch (error) {
-    return c.json(
-      {
-        error:
-          error instanceof Error ? error.message : "No model configured.",
-      },
-      { status: 409 },
-    );
-  }
-
-  const [project, evaluators, prompts, datasets] = await Promise.all([
-    prisma.project.findUnique({
-      where: { id: body.projectId },
-      select: { name: true, language: true, framework: true },
-    }),
-    prisma.evaluator.findMany({
-      where: { projectId: body.projectId },
-      select: { name: true, slug: true, type: true },
-      take: 50,
-    }),
-    prisma.llmPromptConfig.findMany({
-      where: { projectId: body.projectId },
-      select: { handle: true, name: true },
-      take: 50,
-    }),
-    prisma.dataset.findMany({
-      where: { projectId: body.projectId, archivedAt: null },
-      select: { name: true, slug: true },
-      take: 50,
-    }),
-  ]);
-
-  const snapshot = JSON.stringify(
-    { project, evaluators, prompts, datasets },
-    null,
-    2,
-  );
-
-  const stream = streamText({
-    model,
-    system: PROJECT_MEMORY_REFRESH_PROMPT,
-    messages: [
-      {
-        role: "user",
-        content: `Project snapshot (JSON):\n\n${snapshot}`,
-      },
-    ],
-    // Memory-refresh is a one-shot, not a conversation; synthesise a stable
-    // id so the dogfood project can group all refresh traces per project.
-    experimental_telemetry: buildLangyTelemetrySettings({
-      userProjectId: body.projectId,
-      userId: guard.session!.user.id,
-      conversationId: `memory-refresh:${body.projectId}`,
-    }),
-    onFinish: async ({ text }) => {
-      try {
-        const memoryService = LangyProjectMemoryService.create(prisma);
-        await memoryService.writeNewVersion({
-          projectId: body.projectId,
-          content: text,
-          changeReason: "user_refresh",
-          changedById: guard.session!.user.id,
-        });
-        await auditLog({
-          userId: guard.session!.user.id,
-          projectId: body.projectId,
-          action: "langy.project_memory.refresh",
-        });
-      } catch (error) {
-        logger.error({ error }, "failed to persist refreshed project memory");
-      }
-    },
-    onError: (error) => {
-      logger.error({ error }, "project memory refresh stream errored");
-    },
-  });
-
-  return stream.toUIMessageStreamResponse();
-});
-
-// ============================================================================
-// Preferences
-// ============================================================================
-
-app.get("/langy/preferences", async (c) => {
-  const projectId = c.req.query("projectId");
-  const guard = await requireSessionAndPermission(c, projectId);
-  if (guard.error) return guard.error;
-  const service = LangyUserPreferencesService.create(prisma);
-  const prefs = await service.getById({
-    userId: guard.session!.user.id,
-    projectId: projectId!,
-  });
-  return c.json({ preferences: prefs });
-});
-
-app.put("/langy/preferences", async (c) => {
-  const body = (await c.req.json()) as {
-    projectId: string;
-    mode?: "non_expert" | "expert";
-    dismissedSuggestionKinds?: string[];
-  };
-  const guard = await requireSessionAndPermission(c, body.projectId);
-  if (guard.error) return guard.error;
-  const service = LangyUserPreferencesService.create(prisma);
-  let prefs = await service.getById({
-    userId: guard.session!.user.id,
-    projectId: body.projectId,
-  });
-  if (body.mode) {
-    prefs = await service.setMode({
-      userId: guard.session!.user.id,
-      projectId: body.projectId,
-      mode: body.mode,
-    });
-  }
-  if (body.dismissedSuggestionKinds) {
-    prefs = await service.setDismissedSuggestionKinds({
-      userId: guard.session!.user.id,
-      projectId: body.projectId,
-      kinds: body.dismissedSuggestionKinds,
-    });
-  }
-  return c.json({ preferences: prefs });
-});
-
-// ============================================================================
-// Memory clear-all + GDPR export
-// ============================================================================
-
-app.delete("/langy/memory", async (c) => {
-  const projectId = c.req.query("projectId");
-  const guard = await requireSessionAndPermission(c, projectId);
-  if (guard.error) return guard.error;
-  const userId = guard.session!.user.id;
-  const convService = LangyConversationService.create(prisma);
-  const prefService = LangyUserPreferencesService.create(prisma);
-  const result = await convService.clearAllForUser({
-    projectId: projectId!,
-    userId,
-  });
-  await prefService.resetForUser({ projectId: projectId!, userId });
-  await auditLog({
-    userId,
-    projectId: projectId!,
-    action: "langy.memory.clear_all",
-    args: { deletedCount: result.deletedCount },
-  });
-  return c.json({ deletedCount: result.deletedCount });
-});
-
-app.get("/langy/memory/export", async (c) => {
-  const projectId = c.req.query("projectId");
-  const guard = await requireSessionAndPermission(c, projectId);
-  if (guard.error) return guard.error;
-  const userId = guard.session!.user.id;
-  const convService = LangyConversationService.create(prisma);
-  const conversations = await convService.getAll({
-    projectId: projectId!,
-    userId,
-    limit: 1000,
-  });
-  const msgService = LangyMessageService.create(prisma);
-  const conversationsWithMessages = await Promise.all(
-    conversations
-      .filter((c) => c.isOwn)
-      .map(async (c) => ({
-        conversation: c,
-        messages: await msgService.getAllByConversation({
-          conversationId: c.id,
-          projectId: projectId!,
-        }),
-      })),
-  );
-  const prefService = LangyUserPreferencesService.create(prisma);
-  const preferences = await prefService.getById({
-    userId,
-    projectId: projectId!,
-  });
-  await auditLog({
-    userId,
-    projectId: projectId!,
-    action: "langy.memory.export",
-    args: { conversationCount: conversationsWithMessages.length },
-  });
-  return c.json({
-    exportedAt: new Date().toISOString(),
-    projectId,
-    userId,
-    conversations: conversationsWithMessages,
-    preferences,
-  });
-});
+// Conversation CRUD, project-memory CRUD, preferences, and privacy
+// (clear-all + export) live in sibling files. Each register function
+// mutates this same `app`, so middleware applied above (tracer, logger)
+// covers every route uniformly.
+registerLangyConversationRoutes(app);
+registerLangyProjectMemoryRoutes(app);
+registerLangyPreferencesRoutes(app);
+registerLangyPrivacyRoutes(app);
 
 type ParsedState = ReturnType<
   typeof persistedEvaluationsV3StateSchema.parse
